Mon, Jul 31, 2023, 11:18 AM +04:30

+++ CS50x/PSET2/readability/readability.c
@@ -31,10 +31,10 @@ int main(void)
         }
     }
     printf("%i %i %i\n", numC, numW, numS);
-    int L , S = 0;
+    float L , S = 0;
     L = numC/numW*100;
     S = numS/numW*100;
-    printf("%i %i\n", L, S);
+    printf("%f %f\n", L, S);
     float index = 0.0588 * L - 0.296 * S - 15.8;
     if (index >= 16)
     {
